requirements.txt
asgiref==3.8.1
Django==5.0.6
django-crispy-forms==2.1
django-environ==0.11.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
phonenumbers==8.13.38
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
